Fixed login issues in frontend...
