+++ src/main/java/com/crs/controllers/WebController.java
@@ -1,28 +1,72 @@
 package com.crs.controllers;
 
-import com.crs.models.User;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
+import com.crs.controllers.dto.CustomerRegistrationDto;
+import com.crs.services.CustomerService;
+import com.crs.services.UserServiceImpl;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.servlet.ModelAndView;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PostMapping;
+
+import javax.validation.Valid;
 
 @Controller
+@RequiredArgsConstructor
 public class WebController {
 
-    @GetMapping("/login")
-    public String login (){
-        return "login";
+    @GetMapping
+    public String login(){
+        return "home";
     }
 
-    @GetMapping
-    public ModelAndView index () {
+   /* @GetMapping
+    public ModelAndView index() {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         User user = (User) authentication.getPrincipal();
         ModelAndView modelAndView = new ModelAndView();
         modelAndView.setViewName("index");
         modelAndView.addObject(user);
         return modelAndView;
+    }*/
+
+
+    private final UserServiceImpl userService;
+
+    private final CustomerService customerService;
+
+
+    @ModelAttribute("customer")
+    public CustomerRegistrationDto customerRegistrationDto() {
+        return new CustomerRegistrationDto();
+    }
+
+    @GetMapping("login")
+    public String showLoginForm(Model model) {
+        return "login";
+    }
+
+    @GetMapping("registration")
+    public String showRegistrationForm(Model model) {
+        return "registration";
+    }
+
+    @PostMapping("registration")
+    public String registerCustomer(@ModelAttribute("customer") @Valid CustomerRegistrationDto customerRegistrationDto,
+                                   BindingResult result) throws Exception {
+
+        if(userService.doesUserExist(customerRegistrationDto.getEmail())) {
+            result.rejectValue("email", null, "There is already an account registered with that login");
+        }
+
+        if (result.hasErrors()) {
+            return "registration";
+        }
+
+        customerService.createCustomer(customerRegistrationDto);
+        return "login";
     }
