+++ src/main/resources/application.properties
@@ -4,4 +4,4 @@ spring.datasource.password=sekbo2018
 server.port = 9090
 spring.datasource.url=jdbc:mysql://saeed.chcsf7pbopuv.eu-west-2.rds.amazonaws.com/carrentalservice
 spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
-spring.jpa.hibernate.ddl-auto=create-drop
+spring.jpa.hibernate.ddl-auto=update
