+++ src/main/resources/templates/home.html
@@ -1,0 +1,88 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org"
+      xmlns:sec="http://www.thymeleaf.org/thymeleaf-extras-springsecurity3">
+<head>
+    <meta charset="utf-8"/>
+    <meta http-equiv="X-UA-Compatible" content="IE=edge"/>
+    <meta name="viewport" content="width=device-width, initial-scale=1"/>
+
+    <link rel="stylesheet" type="text/css" th:href="@{/webjars/bootstrap/3.3.7/css/bootstrap.min.css}"/>
+    <link href="../static/css/styles.css" th:href="@{/css/styles.css}" rel="stylesheet"/>
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.13.1/css/font-awesome.min.css">
+    <link rel="stylesheet" href="http://maxcdn.bootstrapcdn.com/font-awesome/latest/css/font-awesome.min.css">
+
+
+    <title>Home</title>
+</head>
+<body>
+<div class="container-fluid">
+    <nav class="navbar navbar-dark bg-dark bg-primary">
+        <a class="navbar-brand" href="#">CRS</a>
+        <a href="/" th:href="@{/login}" class="btn btn-primary">Login</a>
+        <a href="/" th:href="@{/registration}" class="btn btn-primary">Register</a>
+    </nav>
+    <div class="container">
+        <h2 class="display-4">
+            <!--            Welcome <span th:text="${user.firstName}"><strong></strong></span>-->
+        </h2>
+        <br/>
+        <br/>
+        <div class="jumbotron">
+            <div>
+                <form class="bookings" action="#">
+                    <input type="text" placeholder="Branches" name="search">
+                    <button type="submit"><i class="fa fa-search"></i></button>
+                </form>
+                <br/>
+            </div>
+        </div>
+
+        <div class="jumbotron">
+            <div>
+                <h3>List of Cars</h3>
+                <h3 class="display-3"><span class="float-right"><a href="/booking/add" class="btn btn-success">Click here</a> </span>
+                </h3>  <!--dropdown for cars -->
+                <br/>
+            </div>
+
+        </div>
+        <div class="jumbotron">
+            <div class="row">
+                <div class="col s12">
+                    <h4>Follow Us</h4>
+                    <p>Follow us on social media for best deals</p>
+                    <a href="#" class="white-text">
+                        <i class="fab fa-facebook fa-4x"></i>
+                    </a>
+                    <a href="#" class="white-text">
+                        <i class="fab fa-twitter fa-4x"></i>
+                    </a>
+                    <a href="#" class="white-text">
+                        <i class="fab fa-linkedin fa-4x"></i>
+                    </a>
+                    <a href="#" class="white-text">
+                        <i class="fab fa-instagram fa-4x"></i>
+                    </a>
+                    <a href="#" class="white-text">
+                        <i class="fab fa-pinterest fa-4x"></i>
+                    </a>
+                    <a href="#" class="white-text">
+                        <i class="fab fa-youtube fa-4x"></i>
+                    </a>
+                </div>
+            </div>
+        </div>
+    </div>
+
+    <!-- Footer -->
+    <footer class="section green white-text center">
+        <p class="flow-text">&copy; CRS OÜ - 2020</p>
+    </footer>
+
+    <script type="text/javascript"
+            th:src="@{/webjars/jquery/3.2.1/jquery.min.js/}"></script>
+    <script type="text/javascript"
+            th:src="@{/webjars/bootstrap/3.3.7/js/bootstrap.min.js}"></script>
+
+</body>
+</html>
