+++ src/main/resources/templates/index.html
@@ -29,13 +29,13 @@
                         Manage DB
                     </a>
                     <div class="dropdown-menu" aria-labelledby="navbarDropdown">
-                        <a class="dropdown-item" href="/cars">Cars</a>
+                        <a class="dropdown-item" href="/car">Cars</a>
                         <div class="dropdown-divider"></div>
-                        <a class="dropdown-item" href="/branches">Branches</a>
+                        <a class="dropdown-item" href="/branch">Branches</a>
                         <div class="dropdown-divider"></div>
-                        <a class="dropdown-item" href="/employees">Employees</a>
+                        <a class="dropdown-item" href="/employee">Employees</a>
                         <div class="dropdown-divider"></div>
-                        <a class="dropdown-item" href="/customers">Customers</a>
+                        <a class="dropdown-item" href="/customer">Customers</a>
                     </div>
                 </li>
                 <li class="nav-item">
